+++ NEWS.md
@@ -1,6 +1,7 @@
 rmarkdown 1.18
 ================================================================================
 
+- `find_external_resources()` works now when multiple files are specified in the `includes` option of the output format (thanks, @andrie, #1677).
 
 
 rmarkdown 1.17